ad-process/src/test/java/com/mydomain/adprocess/marketing/MarketingDepartmentClientTest.java
package com.mydomain.adprocess.marketing;

import org.junit.After;
import org.junit.Before;
import org.junit.Test;
import java.io.ByteArrayInputStream;
import java.io.IOException;
import java.io.ObjectInputStream;
import java.net.ServerSocket;
import java.net.Socket;
import java.util.List;
import static org.junit.Assert.*;

public class MarketingDepartmentClientTest {
    private ServerSocket testServerSocket;
    private final int testPort = 12345; // Use a test port that you know is free
    private Thread serverThread;

    @Before
    public void setUp() throws IOException {
        testServerSocket = new ServerSocket(testPort);
        serverThread = new Thread(() -> {
            try {
                Socket clientSocket = testServerSocket.accept();
                ObjectInputStream ois = new ObjectInputStream(clientSocket.getInputStream());
                @SuppressWarnings("unchecked")
                List<Advertisement> ads = (List<Advertisement>) ois.readObject();
                // Perform assertions or further processing of the received ads
                assertNotNull(ads);
                assertFalse(ads.isEmpty());
                // ... Additional assertions based on your use case ...

                ois.close();
                clientSocket.close();
            } catch (IOException | ClassNotFoundException e) {
                e.printStackTrace();
            }
        });
        serverThread.start();
    }

    @Test
    public void testSendAdvertisements() throws IOException {
        // Redirect System.in for the MarketingDepartmentClient
        System.setIn(new ByteArrayInputStream("1\nAcme Corp\nBuy Acme products!\nPage 3\n1\n100.0\n".getBytes()));
        MarketingDepartmentClient.main(new String[]{"localhost", String.valueOf(testPort)});
        // Main method should connect to testServerSocket and send data
    }

    @After
    public void tearDown() throws IOException {
        testServerSocket.close();
        serverThread.interrupt();
    }
}
